+++ dbx/utils/file_uploader.py
@@ -5,7 +5,6 @@ from typing import Optional, Dict
 import mlflow
 from retry import retry
 
-from dbx.api.configure import ConfigurationManager
 from dbx.api.context import RichExecutionContextClient
 from dbx.utils import dbx_echo
 
@@ -33,9 +32,11 @@ class AbstractFileUploader(ABC):
         if local_file_path in self._uploaded_files:
             remote_path = self._uploaded_files[local_file_path]
         else:
+            dbx_echo(f"Uploading local file {local_file_path}")
             self._upload_file(local_file_path)
             remote_path = "/".join([self._base_uri, str(local_file_path.as_posix())])
             self._uploaded_files[local_file_path] = remote_path
+            dbx_echo(f"Uploading local file {local_file_path} - done")
 
         remote_path = remote_path.replace("dbfs:/", "/dbfs/") if as_fuse else remote_path
         return remote_path
@@ -49,10 +50,8 @@ class MlflowFileUploader(AbstractFileUploader):
     @staticmethod
     @retry(tries=3, delay=1, backoff=0.3)
     def _upload_file(file_path: Path):
-        dbx_echo(f"Uploading file {file_path}")
         posix_path = PurePosixPath(file_path.as_posix())
         parent = str(posix_path.parent) if str(posix_path.parent) != "." else None
-        dbx_echo(f"Uploading file {file_path} - done")
         mlflow.log_artifact(str(file_path), parent)
 
 
@@ -60,19 +59,13 @@ class ContextBasedUploader(AbstractFileUploader):
     def __init__(self, client: RichExecutionContextClient):
         self._client = client
         temp_dir = self._client.get_temp_dir()
-        super().__init__(base_uri=str(temp_dir))
+        super().__init__(base_uri=temp_dir)
 
     def _verify_fuse_support(self):
         dbx_echo("Skipping the FUSE check since context-based uploader is used")
 
-    @staticmethod
-    def _get_current_experiment_id(env_name: str) -> str:
-        info = ConfigurationManager().get(env_name)
-        return mlflow.get_experiment_by_name(info.workspace_dir).experiment_id
-
     def _upload_file(self, local_file_path: Path):
-        temp_remote_file_path = self._client.upload_file(local_file_path, Path(self._base_uri))
-        dbx_echo(f"File uploaded to temp location {temp_remote_file_path}")
+        self._client.upload_file(local_file_path, self._base_uri)
 
     def __del__(self):
         try: